Allow fields to be removed at a specific version (#3417) (#36)
The healthcare API has deprecated a field at beta and removed it from the GA
launch of the api. In order to support this without having to duplicate the
entire resource this change supports defining a single field multiple times
with a different configruation for each version. Specifically it allows a field
to exist as deprecated at one version but not at another.
This also introduces the removed_message field.

Signed-off-by: Modular Magician <magic-modules@google.com>