Removed duplicative flaky test (#4175) (#103)
This test was flaky because the two terraform resources were using the same pubkey, which meant that they ended up with the same remote id (since that is automatically calculated from the pubkey fingerprint) and ran into race conditions where the first test would delete the resource before the second test finished with it.

Additionally, the project id provided was the same as the default project, which meant the setting of the project field was not being tested properly.

Signed-off-by: Modular Magician <magic-modules@google.com>